Seperate enums into individual files and remove unused enum
