+++ app/MonsterConverterInterface/Enums/ChangeAppearanceType.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MonsterConverterInterface.MonsterTypes
+{
+    public enum ChangeAppearanceType
+    {
+        MonsterName,
+        OutfitId,
+        Item
+    }
+}
