+++ app/MonsterConverterInterface/Enums/CombatDamage.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MonsterConverterInterface.MonsterTypes
+{
+    public enum CombatDamage
+    {
+        Physical,
+        Energy,
+        Earth,
+        Fire,
+        LifeDrain,
+        ManaDrain,
+        Healing,
+        Drown,
+        Ice,
+        Holy,
+        Death
+    }
+}
