+++ app/MonsterConverterInterface/Enums/ConditionType.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MonsterConverterInterface.MonsterTypes
+{
+    public enum ConditionType
+    {
+        None,
+        Poison,
+        Fire,
+        Energy,
+        Bleeding,
+        Haste,
+        Paralyze,
+        Outfit,
+        Invisible,
+        Light,
+        ManaShield,
+        InFight,
+        Drunk,
+        ExhaustWeapon, // unused
+        Regeneration,
+        Soul,
+        Drown,
+        Muted,
+        ChannelMutedTicks,
+        YellTicks,
+        Attributes,
+        Freezing,
+        Dazzled,
+        Cursed,
+        ExhaustCombat, // unused
+        Exhaust_heal, // unused
+        Pacified,
+        SpellCoolDown,
+        SpellGroupCoolDown
+    }
+}
