+++ app/MonsterConverterInterface/Enums/Effect.cs
@@ -1,0 +1,99 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MonsterConverterInterface.MonsterTypes
+{
+    public enum Effect
+    {
+        None = 0,
+        DrawBlood,
+        LoseEnergy,
+        Poff,
+        BlockHit,
+        ExplosionArea,
+        ExplosionHit,
+        FireArea,
+        YellowRings,
+        GreenRings,
+        HitArea,
+        Teleport,
+        EnergyHit,
+        MagicBlue,
+        MagicRed,
+        MagicGreen,
+        HitByFire,
+        HitByPoison,
+        MortArea,
+        SoundGreen,
+        SoundRed,
+        PoisonArea,
+        SoundYellow,
+        SoundPurple,
+        SoundBlue,
+        SoundWhite,
+        Bubbles,
+        Craps,
+        GiftWraps,
+        FireworkYellow,
+        FireworkRed,
+        FireworkBlue,
+        Stun,
+        Sleep,
+        WaterCreature,
+        GroundShaker,
+        Hearts,
+        FireAttack,
+        EnergyArea,
+        SmallClouds,
+        HolyDamage,
+        BigClouds,
+        IceArea,
+        IceTornado,
+        IceAttack,
+        Stones,
+        SmallPlants,
+        Carniphila,
+        PurpleEnergy,
+        YellowEnergy,
+        HolyArea,
+        BigPlants,
+        Cake,
+        GiantIce,
+        WaterSplash,
+        PlantAttack,
+        TutorialArrow,
+        TutorialSquare,
+        MirrorHorizontal,
+        MirrorVertical,
+        SkullHorizontal,
+        SkullVertical,
+        Assassin,
+        StepsHorizontal,
+        BloodySteps,
+        StepsVertical,
+        YalahariGhost,
+        Bats,
+        Smoke,
+        Insects,
+        Dragonhead,
+        OrcShaman,
+        OrcShamanFire,
+        Thunder,
+        Ferumbras,
+        ConfettiHorizontal,
+        ConfettiVertical,
+        // 77-157 are empty
+        BlackSmoke = 158,
+        // 159-166 are empty
+        RedSmoke = 167,
+        YellowSmoke = 168,
+        GreenSmoke = 169,
+        PurpleSmoke = 170,
+        EarlyThunder = 171,
+        RagiazBoneCapsule = 172,
+        CriticalDamage = 173,
+        // 174 is empty
+        PlungingFish = 175
+    }
+}
